% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketably.R
\name{market_all}
\alias{market_all}
\title{Make all the functions in a namespace (typically a package)
show ads before running.}
\usage{
market_all(...)
}
\arguments{
\item{...}{Extra arguments passed on to \code{\link{marketably}}}
}
\description{
Call this function in the file zzz.R of a package in order to
advertise for all functions in the package.
}

