% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketably.R
\name{marketably}
\alias{marketably}
\title{Monetize a function by displaying an ad before the function runs}
\usage{
marketably(func, ads, html = FALSE, pause = NULL, impression_url = NULL)
}
\arguments{
\item{func}{Function to add advertising to}

\item{ads}{A character vector from which to pick a random ad}

\item{html}{Whether the ad should be rendered as HTML}

\item{pause}{In seconds, an amount to wait after showing the
ad before running the function.}

\item{impression_url}{A URL to send a GET request to every time
a function is called.}
}
\description{
Monetize a function by displaying an ad before it runs, optionally
with a pause. Can also display HTML, which will appear either in the
RStudio "Viewer" window or (if the viewer window is not available)
in the.
}
\details{
If an \code{impression_url} is given, a GET request is sent along
with the parameters \code{function}, with the function name,
and \code{adId}, with the index of the ad that was selected.
}
\examples{

mean(1:10)

ad_mean <- marketably(mean, "Check out my blog, Variance Explained!")

ad_mean(1:10)

# We can insert a pause first
ad_mean <- marketably(mean, "Check out my blog, Variance Explained!",
                    pause = 3)
ad_mean(1:10)

# or we can include HTML
ad <- paste0("Check out my blog, <a href = 'http://varianceexplained.org'>",
             "Variance Explained</a>!")

ad_mean <- marketably(mean, ad, html = TRUE)
ad_mean(1:10)

}

