% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_all_functions}
\alias{wrap_all_functions}
\title{Wrap all functions in the given namespace}
\usage{
wrap_all_functions(f, env = parent.frame(), ...)
}
\arguments{
\item{f}{Function to wrap every other function with}

\item{env}{Environment to look for functions to wrap}

\item{...}{Extra arguments to pass to the wrapping function}
}
\description{
Wrap all functions in the given namespace with another function.
This is especially useful in a package zzz.R file; for example
as used in the \code{\link{market_all}} function.
}

